+++ src/views/corpora.rs
@@ -7,13 +7,12 @@ use axum::{
     routing::{get, post},
     Form, Router,
 };
-use axum_sessions::extractors::WritableSession;
 use minijinja::context;
 use serde::{Deserialize, Serialize};
 
 use crate::{
     client::corpora,
-    state::{GlobalAppState, SessionArg, SessionState, STATE_KEY},
+    state::{GlobalAppState, Session, SessionArg},
     Result,
 };
 
@@ -29,12 +28,10 @@ struct Corpus {
 }
 
 async fn show(
-    session: WritableSession,
+    session: Session,
     State(app_state): State<Arc<GlobalAppState>>,
 ) -> Result<impl IntoResponse> {
-    let session_state = SessionState::from(&session);
-
-    let selected_corpora = session_state.selected_corpora.clone();
+    let selected_corpora = &session.selected_corpora();
 
     let corpora: Vec<_> = corpora::list(&SessionArg::Session(session.clone()), app_state.as_ref())
         .await?
@@ -50,7 +47,7 @@ async fn show(
         .get_template("corpora.html")?
         .render(context! {
             corpora,
-            session => session_state,
+            session => session,
             filter => "",
         })?;
 
@@ -67,12 +64,10 @@ struct Params {
 }
 
 async fn update(
-    mut session: WritableSession,
+    mut session: Session,
     State(app_state): State<Arc<GlobalAppState>>,
     Form(payload): Form<Params>,
 ) -> Result<impl IntoResponse> {
-    let mut session_state = SessionState::from(&session);
-
     let corpora = corpora::list(&SessionArg::Session(session.clone()), app_state.as_ref()).await?;
     let mut filtered_corpora: Vec<_> = corpora
         .iter()
@@ -81,32 +76,34 @@ async fn update(
         .collect();
     filtered_corpora.sort_by_key(|k| k.to_lowercase());
 
+    let mut selected_corpora = session.selected_corpora().clone();
+
     if let Some(add_corpus) = payload.add_corpus {
-        session_state.selected_corpora.insert(add_corpus);
+        selected_corpora.insert(add_corpus);
     }
 
     if let Some(remove_corpus) = payload.remove_corpus {
-        session_state.selected_corpora.remove(&remove_corpus);
+        selected_corpora.remove(&remove_corpus);
     }
     if payload.add_all_corpora == Some("true".to_string()) {
         // Add all the filtered corpora to the selection
         for c in &filtered_corpora {
-            session_state.selected_corpora.insert(c.clone());
+            selected_corpora.insert(c.clone());
         }
     }
 
     if payload.clear_selection == Some("true".to_string()) {
         // Unselect all corpora
-        session_state.selected_corpora.clear();
+        selected_corpora.clear();
     }
 
     // Update the session
-    session.insert(STATE_KEY, session_state.clone())?;
+    session.set_selected_corpora(selected_corpora)?;
 
     let corpora: Vec<Corpus> = filtered_corpora
         .into_iter()
         .map(|name| Corpus {
-            selected: session_state.selected_corpora.contains(&name),
+            selected: session.selected_corpora().contains(&name),
             name,
         })
         .collect();
@@ -117,7 +114,7 @@ async fn update(
         .render(context! {
             corpora,
             filter => payload.filter,
-            session => session_state,
+            session => session,
         })?;
 
     Ok(Html(html))