+++ src/views/oauth.rs
@@ -1,7 +1,7 @@
 use crate::{
     auth::{schedule_refresh_token, LoginInfo},
     errors::AppError,
-    state::{GlobalAppState, SessionState},
+    state::{GlobalAppState, Session},
     Result,
 };
 use axum::{
@@ -11,7 +11,6 @@ use axum::{
     routing::get,
     Router,
 };
-use axum_sessions::extractors::ReadableSession;
 
 use minijinja::context;
 use oauth2::reqwest::async_http_client;
@@ -52,15 +51,13 @@ async fn redirect_to_login(
 }
 
 async fn logout(
-    session: ReadableSession,
+    session: Session,
     State(app_state): State<Arc<GlobalAppState>>,
 ) -> Result<impl IntoResponse> {
-    let session_state = SessionState::from(&session);
-
-    app_state.login_info.remove(session.id());
+    app_state.login_info.remove(&session.id().to_string());
     let template = app_state.templates.get_template("oauth.html")?;
 
-    let html = template.render(context! {session => session_state})?;
+    let html = template.render(context! {session => session})?;
 
     Ok(Html(html))
 }
@@ -73,16 +70,14 @@ struct CallBackParams {
 }
 
 async fn login_callback(
-    session: ReadableSession,
+    session: Session,
     State(app_state): State<Arc<GlobalAppState>>,
     Query(params): Query<CallBackParams>,
 ) -> Result<impl IntoResponse> {
-    let session_state = SessionState::from(&session);
-
     let template = app_state.templates.get_template("oauth.html")?;
 
     if let Some(error) = params.error {
-        let html = template.render(context! {error, session => session_state})?;
+        let html = template.render(context! {error, session => session})?;
 
         if let Some(state) = params.state {
             app_state.auth_requests.remove(&state);
@@ -104,14 +99,17 @@ async fn login_callback(
                 .await;
             let token = token?;
 
-            let login_info = LoginInfo::new(token.clone(), session.expiry().cloned())?;
+            let login_info = LoginInfo::new(
+                token.clone(),
+                session.expiration_time().map(|t| t.unix_timestamp()),
+            )?;
 
             app_state
                 .login_info
                 .insert(session.id().to_string(), login_info);
 
             let html = template.render(context! {
-                session => session_state,
+                session => session,
             })?;
 
             // Schedule a task that refreshes the token before it expires
@@ -127,8 +125,8 @@ async fn login_callback(
             return Ok((StatusCode::OK, Html(html)));
         }
     }
-    let html = template
-        .render(context! {session => session_state, error => "Empty authorization request."})?;
+    let html =
+        template.render(context! {session => session, error => "Empty authorization request."})?;
     Ok((StatusCode::BAD_REQUEST, Html(html)))
 }